+++ src/sprout/Helpers/DisplayConditions/Acquisition/Referrer.php
@@ -11,16 +11,16 @@
  * For more information, visit <http://getsproutcms.com>.
  */
 
-namespace Sprout\Helpers\DisplayConditions\Session;
+namespace Sprout\Helpers\DisplayConditions\Acquisition;
 
 use Sprout\Helpers\DisplayConditions\DisplayConditionString;
 use Sprout\Helpers\SessionStats;
 
 
 /**
- * Display condtion for the 'utm_source' field
+ * Display condtion for the full HTTP referrer when the session started
  */
-class UtmSource extends DisplayConditionString
+class Referrer extends DisplayConditionString
 {
 
     /**
@@ -31,7 +31,7 @@ class UtmSource extends DisplayConditionString
      */
     protected function getCurrentValue(array $env)
     {
-        return SessionStats::utmSource();
+        return SessionStats::referrer();
     }
 
 }
