+++ src/sprout/Helpers/DisplayConditions/Acquisition/UtmMedium.php
@@ -11,16 +11,16 @@
  * For more information, visit <http://getsproutcms.com>.
  */
 
-namespace Sprout\Helpers\DisplayConditions\Session;
+namespace Sprout\Helpers\DisplayConditions\Acquisition;
 
 use Sprout\Helpers\DisplayConditions\DisplayConditionString;
 use Sprout\Helpers\SessionStats;
 
 
 /**
- * Display condtion for the full HTTP referrer when the session started
+ * Display condtion for the 'utm_medium' field
  */
-class Referrer extends DisplayConditionString
+class UtmMedium extends DisplayConditionString
 {
 
     /**
@@ -31,7 +31,7 @@ class Referrer extends DisplayConditionString
      */
     protected function getCurrentValue(array $env)
     {
-        return SessionStats::referrer();
+        return SessionStats::utmMedium();
     }
 
 }
